--- demos/include/transport_plaintext.h
@@ -1,38 +1,0 @@
-/*
- * Copyright (C) 2020 Amazon.com, Inc. or its affiliates.  All Rights Reserved.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy of
- * this software and associated documentation files (the "Software"), to deal in
- * the Software without restriction, including without limitation the rights to
- * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
- * the Software, and to permit persons to whom the Software is furnished to do so,
- * subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
- * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
- * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
- * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
- * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-#ifndef TRANSPORT_PLAINTEXT_H
-#define TRANSPORT_PLAINTEXT_H
-
-#include <stddef.h>
-#include <stdint.h>
-
-int Transport_ConnectToServer( const char * pServer, uint16_t port );
-
-int32_t Transport_SendExact( int tcpSocket, const void * pMessage, size_t bytesToSend );
-
-int32_t Transport_RecvExact( int tcpSocket, void * pBuffer, size_t bytesToRecv );
-
-int32_t Transport_RecvUpTo( int tcpSocket, void * pBuffer, size_t bytesToRecv );
-
-void Transport_Close( int tcpSocket );
-
-#endif
